Add custom classes to individual footnotes
This small change enables users to add space-seperated classes to a
footnote, using the `class` option in the shortcode.
```
[mfn class="custom-class-1 custom-class-2"]
```

This enables the user to, e.g., style different categories of footnotes
differently. However, the footnotes will still receive their number as
it was before and the classes will be ignored for that matter.

This partially implements my suggestion in issue "Add the possibility to use different categories for footnotes #28", without counting different categories seperately.
However, this might give a better user experience anyway, as each
footnote's id will remain unique.